Implementing a Skiplist in JavaScript
The [skiplist](http://en.wikipedia.org/wiki/Skip_list) is a fun little data-structure that relies on randomness to ensure fast insertion and deletion of elements in a sorted list. [In January](https://twitter.com/mixonic/status/427511311933534208) I wrote a small, understandable version of a skiplist. In this (fast) talk I'll demonstrate and try to help visualize how a skiplist works, then walk through an implementation.