Reframing the Problem
This is a talk about something I have to constantly remind myself and others. Often times we hit a wall and get stuck on a problem, only to realize much later and after lengthy and unnecessary pain the solution was simple *if we just reframed the problem as something else*.

This talk is inspired by, and talks about, one of my favorite "Ah-ha!" moments when getting my degree in CompSci, and that was learning to write a solver for the eight-puzzle (or sliding puzzle). When looked at concretely I thought we would spend all this time writing formulas and deriving some sort of mathematical solution. Instead our AI professor said "If you see the placement of the pieces as a Vertex and a move as an Edge, you've reframed and reduced the problem to something already solved: finding the shortest path between two Vertexes.