Make awesome GIFs using the ncolorpalette
The ncolorpalette is a single page that applies a new, limited palette to a full-color image. It uses k-means clustering to determine how to apply the palette. I'd like to show how that works, as well as demonstrate how nearly the same code is used as a revisit.link service.

- http://kirbysayshi.com/ncolorpalette/
- hub.revisit.link (Gameboy palette)
- https://github.com/kirbysayshi/revisit-ncolorpalette-service