October talk proposal
Twitter flame wars. Open source project failures through contributor fallout. Bullying that leads to the bullying of bullies.

How do we maintain our communities? How do we grow them? How is anyone in their right mind stepping into the caustic social media environment of programming we have created for ourselves?

There are steps we can take with user groups, conferences, online worlds that can help support an inclusive environment that allows for debate, learning, growing, and excitement. The social sciences have 100 years of research that can provide us with tools to treat our fellow programmers with care. Much of this has been battle-tested in the organizations we run. The bubbles we live in. I'll share how we're creating a better world that is welcoming to programmers, new and experienced, so that we're leading by example. 

Oh and try not to soapbox because that's gross.

Your words are powerful, and “actions speak louder than words, but not nearly as often.”