Builds apps, not infrastructure
While most of us love building apps in JavaScript, we'd also love it if people
used them, and if we're lucky, *lots* of people. Problem is, then we really
need to care about about how our app runs and scales, and it's best to
think about that before that moment of fame arrives.

I'll show a few tools and services that can help you manage your
infrastructure, and the pros/cons of different approaches, so you can be
prepared and spend more time building your apps.