Update index.html
@brianloveswords you made me do this

There have been a fair amount of talks on hardware already I'm guessing (oh hai I'm new to this meetup), but I wanted to provide a slightly different angle, which is writing libraries for hardware rather than just using them. There are a lot of head-scratching moments when you begin, so a high level guide with some tibits to get the audience started on bare-metal bits and bytes via Javascript would be my aim. Presented in a highly entertaining, non-dry format of course.