Update index.html
When I've been able to get a ticket, I've seen an increasing number of talks about React.js and the technologies around the Component Lifecycle, Server-side rendering and Flux. 

I'd like to talk here about the nitty gritty of moving a codebase from "not-react", Backbone perhaps, to React in an incremental way. In doing so, I'll also introduce some strong points of React(hint, it isn't the diffing and some parts are non-technical, even) surrounding that provide compelling reasons for a business to consider an incremental migration.