Dynamic Forms on the Fly
Creating forms on the fly with React and JSON Schema.

With a few configuration options and custom validation errors, you can create forms on the fly with React.JS and JSON Schema. Forms should be, in most cases, derived from data, whether that is a model with empty attributes for a POST or with pre-set attributes for a PUT.  From there React can take over and render the form, validate, display errors and finally submit. 

Extra bonus(or extra points off?!?): Will try to get a view helper for Rails written to allow this to be rendered server-side.