Adds Aaron Petcoff's speaker proposal
In this talk I will describe a project I've begun working in cooperation with students and faculty from the University of Michigan-Dearborn to visually represent data documenting housing speculation in Detroit. This project aims to engage students, assist community organizers, and educate the public on the housing crisis in Detroit.