Talk for JavaScript Bridge in iOS webviews
I want to talk a bit about the JavaScript Bridge, an iOS tool to run native code through JavaScript.  In the last two versions of iOS (8 and 9), Apple has been giving more power to JavaScript. I think I can find 9 other things that are interesting to go along with it.