Talk Proposal: Click, Clack, Code: Mechanical Keyboards & You
This talk is an intro to mechanical keyboards for people who type for a living.
I will introduce the vocabulary around mechanical keyboards (and keyboards in
general), the options available to the professional typist, and the
considerations in choosing your typing technology. I will also try to have some
props available for sample at bevera.js.