Aesthetics of Flux
What is beautiful code?  I believe the answer is code that expressively employs the power of architecture to clearly communicate with the reader, leaving minimal room for confusion or misunderstanding.  What is beautiful in C will be different than what is beautiful in Smalltalk, and both will be different than beautiful Haskell code.

Often the language itself imposes the paradigm, but Javascript is so easy-going we have to decide for ourselves what type of code we want to create.  Despite the extra work this makes for us, it also allows for innovations that would go against the grain in any other language.  We can write code that is beautiful the way functional code is beautiful, or code that is beautiful the way object-oriented code is beautiful, but it also leaves the door open to novel forms.

One such new approach is Flux, an architecture pattern that emerged from Facebook.  It is an example of reactive programming, and it can be a mind twist for people used to object-oriented or functional paradigms.  Unfortunately the documentation is expansive and at times obtuse, diving quickly into the weeds.  I offer instead a high-level explanation, involving fewer justifications and more pictures to demonstrate how Flux fits the ways JavaScript itself is structured, and how it supports writing beautiful JavaScript code.