iBeacon Your Pardon?
The Sloppy Moose Running Club in Sacramento had an event check-in problem. Paper sign-in forms in small spaces caused long lines and a frustrating experience. This talk shows various ways that iBeacon technology can be used (and abused) to interact with a user's relative location which ultimately led to a quicker and more accurate check-in experience for the Moosers.