[April] Propose "webcoder" talk
I was hoping to give a talk about a project I'm working on -- webcoder -- which allows you to edit code remotely in a browser (github.com/mjlyons/webcoder, screenshot attached). Unlike Sublime, Atom, etc., which assume the editor runs on the same computer as the code, webcoder lets you work from any computer's web browser regardless of where the code lives/runs on. So you could store and run your code on a a Digital Ocean instance and and edit the code from any computer that can run a web browser. It's certainly not as full featured as pretty much any other code editor, but it does syntax highlighting for a bunch of languages (including JS), has a folder navigator, and Ctrl-P fast file search. It's good enough that I use it for my day job, and it's written in ES6 with React for the client and Express for the server.

Thanks,
Mike