How to Print Your Own ASCII Art Selfies
*How to Print Your Own ASCII Art Selfies*

Everyone loves selfies and ASCII art. In this talk I will live code a simple Node.js app that will allow the audience to send a selfie via MMS and have it printed in ASCII art with a thermal receipt printer. The printer has bluetooth capabilities and is the same device that the meter maids in NYC use.