Stop, Collaborate, and Listen!
When it comes to team process there are a million variations of Agile, Scrum, Lean, etc. that engineering teams use to be productive and build good products. But no matter what process your team uses, how can you be the best dev you can be? By collaborating, of course! In this talk, we'll go over some ways to apply engineering skills to collaborate better with designers and product people on your team.