Bublé
So a couple of weeks ago I released a Babel alternative called [Bublé](https://buble.surge.sh). It has some very different design goals and constraints to Babel, which might be interesting to explore.

Mostly I'd like to talk about the approach to code transformation. Typically, the way that compilers work (as far as I can tell – certainly in Babel's case) is to generate an AST from the source code, analyse that AST, manipulate it, then generate code from it. Bublé skips the last two steps, modifying the code in place instead. This approach is *waaaay* faster, and doesn't bork up your formatting and code style. And it's done in a way that you still get sourcemaps etc.

Unless you think that's too in-the-weeds in which case there are a few other angles:

* ES2015 footguns
* pragmatic vs ideological programming
* you say NIH, I say progress
* designing APIs around humans, rather than the implementation

Cheers!