"fifty points at random"
When does process become art? When do tools become collaborators? Let's explore the world of procedural art -- both analog and digital -- through the wall drawing of Sol LeWitt. This talk teaches how to make art in the style of LeWitt's wall drawings with CSS and JavaScript, and along the way provides some insight into the history of procedural art, the nature of randomness, and how tools and algorithms for creating art lead to new forms of collaboration.