Anatomy of a client side ad fraud
In December we came across a massive ad fraud which was cloaking the real domain of a web page using client side techniques:

You can see more about it here:
https://github.com/fertel/cloak-pub

Here is a press release from when Comscore found it:
http://www.comscore.com/Insights/Press-Releases/2016/4/comScore-Labs-Discovers-Placement-Laundering-Attack-in-SafeFrame

During the talk I will go through the way in which they perpetrate the fraud. How they go about doing it by overriding some of Javascript's standard built in objects and functions. How one can/can't protect against this - and why it's hard to trust someone else's windows and scripts.