It's about Time
No seriously, it's about time. The short description is I want to rant for 10 minutes on time and how it's a BS social construct. The long description is I want to talk about the complexities of time handling in applications. This is not a talk that ends with answers but more like it ends with a checklist and a reminder to _really_ spend 3x the time you normally would have really cementing and thinking about anything related to time in your applications.