Testing in parallel with AVA
[AVA](https://github.com/avajs/ava) is a JavaScript test runner that runs your tests concurrently and your files in parallel, giving you maximum speed (especially when dealing with IO!). Other features include built-in Babel (including `async`/`await` functions), test macros, watcher, TAP output, great assertion messages through [power-assert](https://github.com/power-assert-js/power-assert).

My talk would be an introduction to AVA and how it can benefit your React and Node testing with its speed, developer experience, and isolated processes. There'll be a lot of code and theory to explain why and how concurrency helps and how atomic tests make your life easier. Lastly, I want to talk about managing and inclusive and inviting open source project like AVA and my experience with it.