Add July proposal
In May, I released a project called Property Praxis <http://propertypraxis.org> as part of a small critical research project into Detroit property ownership. Through our research we found that one out of every five parcels in Detroit were owned by speculators. Beyond radically shaping Detroit's geographic and physical space, this has a profound impact on housing in Detroit–a majority African American city that was among the most affected areas during the 2008 economic crisis. Property Praxis is the result of a year of research into property ownership in Detroit. We use an interactive map built on CartoDB, Ember, and Leaflet to visualize the result of our research. Our hope is that this map can be used as a tool to increase transparency in Detroit's property market, inform policy makers and media creators, and helps support community activists fighting against housing injustice in Detroit and elsewhere.