Emoji-speaking Chatbots: building a Bot using Twilio and Node.js
Over the past six month, two designers and I built an SMS-based bot that facilitates games of Emoji pictionary between friends. How it works: Player A gets a phrase and texts out some emoji. Players B, C, and D try to guess what the emojis mean. Cue trash talking and verbal abuse.

Over the course of building the bot, we've learned a few things that might be of interest to fellow Javascripters:

- Handling emojis and dealing with the astral plane in Node.js
- Design-focused best practices for communication via text with bots
- Testing strategies for keeping bots sane
- Strategies for keeping costs down over SMS