Proposing peer design talk
Designing in a vacuum is challenging. It’s more than challenging - it’s hard, painful, sad, depressing, defeating, pointless, infuriating, lonely - you get my point? It can potentially be debilitating for a creative to be working in a silo, which from time to time could happen on a project. To address this, I've been experimenting with some techniques to support pair designing that I'd like to share.