"ECMAScript #FreeJewelry" You Can Wear Today!
# "ECMAScript #FreeJewelry"...You Can Wear Today!

Ecmascript is providing us with an assortment of new features to the Javascript language. Asyncs, Promises, Generators, Type systems and other deeper features. We can use them today but are a nice to have not a must have.

In this talk we will focus on a handful of new ECMAScript features you can use in the browser today! And if you act now you’ll get string templates, object literal defaults, destructors,
variable swapping for free!