Update index.html
NESpectre is a genuine 8-bit NES console that will be supernaturally influenced by YOU and the rest of the audience.
http://imgur.com/a/oUqEk

The project involves a dual-port RAM circuit that lets you access the entire 2KB of memory in an NES via a serial port and Node.js / socket.io to serve a user interface to devices. The audience can then modify the ram using a (spooky) currency to change the fate of the unsuspecting player. The main demo runs a 2 player game of Contra and lets the audience add/remove lives and weapons, corrupt graphics and sound, and a few other bizarre effects we found when exploring the game. 

If there's time we'll run a Track & Field demo where the audience is split into teams and the contribution of the audience's taps will make each player run. The results are usually hilarious since we push the speed beyond what's possible by a controller input and the long jump gets really long (off the top of the screen and wrapping around a few times before he lands).

Audience view / UI :
http://imgur.com/a/7vhQY

Technical details :
http://www.andrewreitano.com/2907125-nes-dual-port-ram-interface