Code Stories: A Year of Technical Blogging
Blogs: we all love reading them and we all dread writing them. Technical blogging is one of the easiest ways to give back to the tech community, and it was the place that I started when I was fresh graduate of a coding bootcamp and had just started my first developer job.

But we all dread it for a reason: blogging is intimidating. It’s hard to feel like your ideas are valid, and to actually write something without doubting your own qualifications, let alone whether anyone is actually reading anything that you write. So how do we reconcile this with the fact that blogging is one of the best ways to level up your skills?

In this talk, I’ll share what I learned in my one-year project of writing a technical blog every single week. In researching and writing my series of “Technical Tuesday” blog posts (http://vaidehijoshi.github.io), I realized not only what makes blogging so hard, but also how to develop your own technical writing style and structure. We’ll look at the different types of technical blogs that are out there, what the most successful blog posts do well, and how to that pinpoint and then imitate that style in our own technical writing.