Sam Gimbel submits his first talk!
"Helping people" has always been my MO, but how do you do that without getting stuck in systems that perpetuate badness? In this talk, I'll take you on a journey through medicine, tech, and education to highlight how responsibly "nudging" systems in the right direction can be surprisingly effective. It's an evidence-based journey, a journey I've lived, and one that speaks to one of the prime goals of most of us: making a difference with the work we do.