Added myself as first speaker
I'm in town on the 18th, so I thought it might be a good idea to meet some friends and speak at the event.

## abstract

At @nearForm, we like to scratch our own itch, and we build tools based on the needs of our customers and our own needs as consultants. @davidmarkclements and myself were in London conducting a performance training and consulting, and we noticed one notable pattern: the routes that were logging more were the slowest ones. Moreover, disabling logging allowed far greater throughput. We started a quest to assess the cost of logging. In the end, we ended writing down our own logger, https://www.npmjs.com/package/pino. This talk is the story of our journey.

# Seen at

http://www.nearform.com/nodecrunch/cost-logging/
https://vimeo.com/180476148
https://www.youtube.com/watch?v=zja-_IYNrFc