Added myself to speaker list
Working talks titles:
1. Let's not romanticize learning to code. It's challenging: mentally, emotionally and physically.
2. learningToCode == learningAboutYourself

I’m a self-taught developer, and the process itself was an exercise in learning about myself. At points I was miserable, experiencing bouts of depression and angst, and quitting multiple times. Following the last time I ‘quit learning to program’ I realized there was some part of me that was self-sabotaging, and I decided to quit that instead. So I left my job, got an internship as a junior developer and have been programming full-time since. 

Part of my angst came from this sense that programming was somehow this innate skill I needed to have been born with, or should be able to develop effortlessly. I now feel that’s insane. But, the reality is, there are other people who think the same and the media projects this idea of learning to program / being a programmer as some sort of romance story: You read an article about programming. You take an online class and you’re instantly hooked. It’s meant to be, happiness for everyone! NOPE.

I feel that we should be sincere and share what the reality of learning to program is really like, regardless of your path. No one experiences it the same way, and sharing mainly the stories of those who were fortunate to excel as programmers, can set others up for failure (or make them feel like they’re failing). Let’s stop romanticizing learning to program / being a programmer - it’s hard, but worth it, if it’s the right fit for you.